Fixed support of multiple sites
The new version of a hack is somehow different. Though, it attempts to
do the same trick altering the underlying connection, it does a
reconnect instead of just patching the URL. As a result, this allows
different resources to work with distinct `site`s at the same time.

See [related upstream issue](
  https://github.com/JsonApiClient/json_api_client/issues/215
) for details.